Fix to #7364 - `Distinct()` after `Select()`ing a property doesn't generate correct SQL if query also contains optional navigation
Problem was that optional navigation is translated to GroupJoin which in turn forces materialization on entire entity. This causes Distinct to be applied on all columns of that entity, even if the customer specified a subset of columns to perform Distinct on.
Fix is to always perform Distinct on the client if the query also contains GroupJoin